+++ core/src/executor/datafusion/iceberg_file_task_scan.rs
@@ -38,7 +38,7 @@ use iceberg::scan::FileScanTask;
 use iceberg_datafusion::physical_plan::expr_to_predicate::convert_filters_to_predicate;
 use iceberg_datafusion::to_datafusion_error;
 
-use super::SEQ_NUM;
+use super::datafusion_processor::SEQ_NUM;
 
 /// An execution plan for scanning iceberg file scan tasks
 #[derive(Debug)]
