+++ core/src/executor/datafusion/mod.rs
@@ -19,17 +19,19 @@ use ::datafusion::{
     prelude::{SessionConfig, SessionContext},
 };
 use async_trait::async_trait;
-use file_scan_task_table_provider::IcebergFileScanTaskTableProvider;
+use datafusion_processor::{DataFusionTaskContext, DatafusionProcessor};
 use futures::{StreamExt, future::try_join_all};
 use iceberg::{
-    arrow::schema_to_arrow_schema,
     io::FileIO,
-    scan::FileScanTask,
-    spec::{DataFile, NestedField, PartitionSpec, PrimitiveType, Schema, Type},
+    spec::{DataFile, PartitionSpec, Schema},
     writer::{
-        base_writer::data_file_writer::{ DataFileWriterBuilder}, file_writer::{
-            location_generator::{DefaultFileNameGenerator, DefaultLocationGenerator}, ParquetWriterBuilder
-        }, function_writer::fanout_partition_writer::{FanoutPartitionWriterBuilder}, IcebergWriter, IcebergWriterBuilder
+        IcebergWriter, IcebergWriterBuilder,
+        base_writer::data_file_writer::DataFileWriterBuilder,
+        file_writer::{
+            ParquetWriterBuilder,
+            location_generator::{DefaultFileNameGenerator, DefaultLocationGenerator},
+        },
+        function_writer::fanout_partition_writer::FanoutPartitionWriterBuilder,
     },
 };
 use sqlx::types::Uuid;
@@ -39,16 +41,10 @@ use tokio::task::JoinHandle;
 use crate::{CompactionConfig, CompactionError};
 
 use super::{CompactionExecutor, CompactionResult, InputFileScanTasks, RewriteFilesStat};
+pub mod datafusion_processor;
 pub mod file_scan_task_table_provider;
 pub mod iceberg_file_task_scan;
-pub mod sql_builder;
 
-const SEQ_NUM: &str = "seq_num";
-const FILE_PATH: &str = "file_path";
-const POS: &str = "pos";
-const DATA_FILE_TABLE: &str = "data_file_table";
-const POSITION_DELETE_TABLE: &str = "position_delete_table";
-const EQUALITY_DELETE_TABLE: &str = "equality_delete_table";
 const DEFAULT_PREFIX: &str = "10";
 
 pub struct DataFusionExecutor {}
@@ -71,7 +67,7 @@ impl CompactionExecutor for DataFusionExecutor {
             .unwrap_or(DEFAULT_PREFIX.to_owned());
         let mut session_config = SessionConfig::new();
         session_config = session_config.with_target_partitions(target_partitions);
-        let ctx = SessionContext::new_with_config(session_config);
+        let ctx = Arc::new(SessionContext::new_with_config(session_config));
 
         let mut stat = RewriteFilesStat::default();
         let rewritten_files_count = input_file_scan_tasks.input_files_count();
@@ -82,64 +78,18 @@ impl CompactionExecutor for DataFusionExecutor {
             equality_delete_files,
         } = input_file_scan_tasks;
 
-        let mut data_fusion_task_context = DataFusionTaskContext::builder()?
+        let datafusion_task_ctx = DataFusionTaskContext::builder()?
             .with_schema(schema)
             .with_datafile(data_files)
             .with_position_delete_files(position_delete_files)
             .with_equality_delete_files(equality_delete_files)
-            .build()?;
+            .build_merge_on_read()?;
 
-        let need_seq_num = data_fusion_task_context.need_seq_num();
-        let need_file_path_and_pos = data_fusion_task_context.need_file_path_and_pos();
-
-        let data_file_schema = data_fusion_task_context.data_file_schema.take().unwrap();
-        let input_schema = data_fusion_task_context.input_schema.take().unwrap();
-
-        // register data file table provider
-        Self::register_data_table_provider(
-            &data_file_schema,
-            data_fusion_task_context.data_files.take().unwrap(),
-            file_io.clone(),
-            &ctx,
-            DATA_FILE_TABLE,
-            need_seq_num,
-            need_file_path_and_pos,
-            batch_parallelism,
-        )?;
-
-        if let Some(position_delete_schema) = data_fusion_task_context.position_delete_schema.take()
-        {
-            Self::register_delete_table_provider(
-                &position_delete_schema,
-                data_fusion_task_context
-                    .position_delete_files
-                    .take()
-                    .unwrap(),
-                file_io.clone(),
-                &ctx,
-                POSITION_DELETE_TABLE,
-                batch_parallelism,
-            )?;
-        }
-
-        if let Some(equality_delete_schema) = data_fusion_task_context.equality_delete_schema.take()
-        {
-            Self::register_delete_table_provider(
-                &equality_delete_schema,
-                data_fusion_task_context
-                    .equality_delete_files
-                    .take()
-                    .unwrap(),
-                file_io.clone(),
-                &ctx,
-                EQUALITY_DELETE_TABLE,
-                batch_parallelism,
-            )?;
-        }
-
+        let (df, input_schema) =
+            DatafusionProcessor::new(ctx, datafusion_task_ctx, batch_parallelism, file_io.clone())
+                .execute()
+                .await?;
         let arc_input_schema = Arc::new(input_schema);
-
-        let df = ctx.sql(&data_fusion_task_context.merge_on_read_sql).await?;
         let batchs = df.execute_stream_partitioned().await?;
         let mut futures = Vec::with_capacity(batch_parallelism);
         // build iceberg writer for each partition
@@ -152,8 +102,14 @@ impl CompactionExecutor for DataFusionExecutor {
             let future: JoinHandle<
                 std::result::Result<Vec<iceberg::spec::DataFile>, CompactionError>,
             > = tokio::spawn(async move {
-                let mut data_file_writer =
-                    Self::build_iceberg_writer(data_file_prefix, dir_path, schema, file_io,partition_spec).await?;
+                let mut data_file_writer = Self::build_iceberg_writer(
+                    data_file_prefix,
+                    dir_path,
+                    schema,
+                    file_io,
+                    partition_spec,
+                )
+                .await?;
                 while let Some(b) = batch.as_mut().next().await {
                     data_file_writer.write(b?).await?;
                 }
@@ -176,7 +132,7 @@ impl CompactionExecutor for DataFusionExecutor {
             .iter()
             .map(|f| f.file_size_in_bytes())
             .sum();
-        stat.rewritten_files_count = rewritten_files_count as u32;
+        stat.rewritten_files_count = rewritten_files_count;
 
         Ok(CompactionResult {
             data_files: output_data_files,
@@ -186,85 +142,13 @@ impl CompactionExecutor for DataFusionExecutor {
 }
 
 impl DataFusionExecutor {
-    #[allow(clippy::too_many_arguments)]
-    fn register_data_table_provider(
-        schema: &Schema,
-        file_scan_tasks: Vec<FileScanTask>,
-        file_io: FileIO,
-        ctx: &SessionContext,
-        table_name: &str,
-        need_seq_num: bool,
-        need_file_path_and_pos: bool,
-        batch_parallelism: usize,
-    ) -> Result<(), CompactionError> {
-        Self::register_table_provider_impl(
-            schema,
-            file_scan_tasks,
-            file_io,
-            ctx,
-            table_name,
-            need_seq_num,
-            need_file_path_and_pos,
-            batch_parallelism,
-        )
-    }
-
-    fn register_delete_table_provider(
-        schema: &Schema,
-        file_scan_tasks: Vec<FileScanTask>,
-        file_io: FileIO,
-        ctx: &SessionContext,
-        table_name: &str,
-        batch_parallelism: usize,
-    ) -> Result<(), CompactionError> {
-        Self::register_table_provider_impl(
-            schema,
-            file_scan_tasks,
-            file_io,
-            ctx,
-            table_name,
-            false,
-            false,
-            batch_parallelism,
-        )
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn register_table_provider_impl(
-        schema: &Schema,
-        file_scan_tasks: Vec<FileScanTask>,
-        file_io: FileIO,
-        ctx: &SessionContext,
-        table_name: &str,
-        need_seq_num: bool,
-        need_file_path_and_pos: bool,
-        batch_parallelism: usize,
-    ) -> Result<(), CompactionError> {
-        let schema = schema_to_arrow_schema(schema)?;
-        let data_file_table_provider = IcebergFileScanTaskTableProvider::new(
-            file_scan_tasks,
-            Arc::new(schema),
-            file_io,
-            need_seq_num,
-            need_file_path_and_pos,
-            batch_parallelism,
-        );
-
-        ctx.register_table(table_name, Arc::new(data_file_table_provider))
-            .unwrap();
-        Ok(())
-    }
-
     async fn build_iceberg_writer(
         data_file_prefix: String,
         dir_path: String,
         schema: Arc<Schema>,
         file_io: FileIO,
         partition_spec: Arc<PartitionSpec>,
-    ) -> Result<
-        Box<dyn IcebergWriter>,
-        CompactionError,
-    > {
+    ) -> Result<Box<dyn IcebergWriter>, CompactionError> {
         let location_generator = DefaultLocationGenerator { dir_path };
         let unique_uuid_suffix = Uuid::now_v7();
         let file_name_generator = DefaultFileNameGenerator::new(
@@ -282,212 +166,20 @@ impl DataFusionExecutor {
         );
         let data_file_builder =
             DataFileWriterBuilder::new(parquet_writer_builder, None, partition_spec.spec_id());
-            let iceberg_output_writer =if partition_spec.fields().is_empty() {
-                Box::new(data_file_builder.build().await?) as Box<dyn IcebergWriter>
-            } else {
-                Box::new(FanoutPartitionWriterBuilder::new(
-                                    data_file_builder,
-                                    partition_spec.clone(), 
-                                    schema,
-                                )?.build().await?) as Box<dyn IcebergWriter>
-            };
+        let iceberg_output_writer = if partition_spec.fields().is_empty() {
+            Box::new(data_file_builder.build().await?) as Box<dyn IcebergWriter>
+        } else {
+            Box::new(
+                FanoutPartitionWriterBuilder::new(
+                    data_file_builder,
+                    partition_spec.clone(),
+                    schema,
+                )?
+                .build()
+                .await?,
+            ) as Box<dyn IcebergWriter>
+        };
         Ok(iceberg_output_writer)
-    }
-}
-
-struct DataFusionTaskContext {
-    pub(crate) data_file_schema: Option<Schema>,
-    pub(crate) input_schema: Option<Schema>,
-    pub(crate) data_files: Option<Vec<FileScanTask>>,
-    pub(crate) position_delete_files: Option<Vec<FileScanTask>>,
-    pub(crate) equality_delete_files: Option<Vec<FileScanTask>>,
-    pub(crate) position_delete_schema: Option<Schema>,
-    pub(crate) equality_delete_schema: Option<Schema>,
-    pub(crate) merge_on_read_sql: String,
-}
-
-struct DataFusionTaskContextBuilder {
-    schema: Arc<Schema>,
-    data_files: Vec<FileScanTask>,
-    position_delete_files: Vec<FileScanTask>,
-    equality_delete_files: Vec<FileScanTask>,
-}
-
-impl DataFusionTaskContextBuilder {
-    pub fn with_schema(mut self, schema: Arc<Schema>) -> Self {
-        self.schema = schema;
-        self
-    }
-
-    pub fn with_datafile(mut self, data_files: Vec<FileScanTask>) -> Self {
-        self.data_files = data_files;
-        self
-    }
-
-    pub fn with_position_delete_files(mut self, position_delete_files: Vec<FileScanTask>) -> Self {
-        self.position_delete_files = position_delete_files;
-        self
-    }
-
-    pub fn with_equality_delete_files(mut self, equality_delete_files: Vec<FileScanTask>) -> Self {
-        self.equality_delete_files = equality_delete_files;
-        self
-    }
-
-    // build data fusion task context
-    pub fn build(self) -> Result<DataFusionTaskContext, CompactionError> {
-        let highest_field_id = self.schema.highest_field_id();
-        // Build scheam for position delete file, file_path + pos
-        let position_delete_schema = Schema::builder()
-            .with_fields(vec![
-                Arc::new(NestedField::new(
-                    1,
-                    FILE_PATH,
-                    Type::Primitive(PrimitiveType::String),
-                    true,
-                )),
-                Arc::new(NestedField::new(
-                    2,
-                    POS,
-                    Type::Primitive(PrimitiveType::Long),
-                    true,
-                )),
-            ])
-            .build()?;
-        // Build schema for equality delete file, equality_ids + seq_num
-        let mut equality_ids: Option<Vec<i32>> = None;
-        for i in &self.equality_delete_files {
-            if let Some(ids) = equality_ids.as_ref() {
-                if ids.eq(&i.equality_ids) {
-                    continue;
-                } else {
-                    return Err(CompactionError::Config("equality_ids not equal".to_owned()));
-                }
-            } else {
-                equality_ids = Some(i.equality_ids.clone());
-            }
-        }
-        let mut equality_delete_vec = vec![];
-        if let Some(ids) = equality_ids.as_ref() {
-            for i in ids {
-                let field = self
-                    .schema
-                    .field_by_id(*i)
-                    .ok_or_else(|| CompactionError::Config("equality_ids not found".to_owned()))?;
-                equality_delete_vec.push(field.clone());
-            }
-        }
-        let equality_join_names: Vec<_> =
-            equality_delete_vec.iter().map(|i| i.name.clone()).collect();
-
-        if !equality_delete_vec.is_empty() {
-            equality_delete_vec.push(Arc::new(NestedField::new(
-                highest_field_id + 1,
-                SEQ_NUM,
-                Type::Primitive(PrimitiveType::Long),
-                true,
-            )));
-        }
-        let equality_delete_schema = Schema::builder().with_fields(equality_delete_vec).build()?;
-        let need_file_path_and_pos = !self.position_delete_files.is_empty();
-        let need_seq_num = !equality_join_names.is_empty();
-
-        // Build schema for data file, old schema + seq_num + file_path + pos
-        let project_names: Vec<_> = self
-            .schema
-            .as_struct()
-            .fields()
-            .iter()
-            .map(|i| i.name.clone())
-            .collect();
-        let highest_field_id = self.schema.highest_field_id();
-        let mut add_schema_fields = vec![];
-        // add sequence number column if needed
-        if need_seq_num {
-            add_schema_fields.push(Arc::new(NestedField::new(
-                highest_field_id + 1,
-                SEQ_NUM,
-                Type::Primitive(PrimitiveType::Long),
-                true,
-            )));
-        }
-        // add file path and position column if needed
-        if need_file_path_and_pos {
-            add_schema_fields.push(Arc::new(NestedField::new(
-                highest_field_id + 2,
-                FILE_PATH,
-                Type::Primitive(PrimitiveType::String),
-                true,
-            )));
-            add_schema_fields.push(Arc::new(NestedField::new(
-                highest_field_id + 3,
-                POS,
-                Type::Primitive(PrimitiveType::Long),
-                true,
-            )));
-        }
-        // data file schema is old schema + seq_num + file_path + pos. used for data file table provider
-        let data_file_schema = self
-            .schema
-            .as_ref()
-            .clone()
-            .into_builder()
-            .with_fields(add_schema_fields)
-            .build()?;
-        // input schema is old schema. used for data file writer
-        let input_schema = self.schema.as_ref().clone();
-
-        let sql_builder = sql_builder::SqlBuilder::new(
-            &project_names,
-            &self.position_delete_files,
-            &self.equality_delete_files,
-            &equality_join_names,
-            need_seq_num,
-            need_file_path_and_pos,
-        );
-        let merge_on_read_sql = sql_builder.build_merge_on_read_sql();
-
-        Ok(DataFusionTaskContext {
-            data_file_schema: Some(data_file_schema),
-            input_schema: Some(input_schema),
-            data_files: Some(self.data_files),
-            position_delete_files: Some(self.position_delete_files),
-            equality_delete_files: Some(self.equality_delete_files),
-            position_delete_schema: if need_file_path_and_pos {
-                Some(position_delete_schema)
-            } else {
-                None
-            },
-            equality_delete_schema: if need_seq_num {
-                Some(equality_delete_schema)
-            } else {
-                None
-            },
-            merge_on_read_sql,
-        })
-    }
-}
-
-impl DataFusionTaskContext {
-    pub fn builder() -> Result<DataFusionTaskContextBuilder, CompactionError> {
-        Ok(DataFusionTaskContextBuilder {
-            schema: Arc::new(Schema::builder().build()?),
-            data_files: vec![],
-            position_delete_files: vec![],
-            equality_delete_files: vec![],
-        })
-    }
-
-    pub fn need_file_path_and_pos(&self) -> bool {
-        self.position_delete_files
-            .as_ref()
-            .is_some_and(|v| !v.is_empty())
-    }
-
-    pub fn need_seq_num(&self) -> bool {
-        self.equality_delete_files
-            .as_ref()
-            .is_some_and(|v| !v.is_empty())
     }
 }
